+++ tests/io/block_manager_test.cpp
@@ -65,8 +65,8 @@ TEST(BlockManager, Test1) {
 
     for (size_t i = 0; i < block_type::size; ++i)
     {
-        block[0].elem_[i].integer = i;
-        block[1].elem_[i].integer = i;
+        block[0].elem_[i].integer = static_cast<int>(i);
+        block[1].elem_[i].integer = static_cast<int>(i);
     }
     for (size_t i = 0; i < nblocks; ++i)
         reqs[i] = block[i].write(bids[i], my_handler());
